More popups for process info; refactoring
